Move `FuzzProject` out to its own module
Just code motion.
